feat: Logout redirects to login page; Login from plans redirects user back to plans page
